requirements.txt
python-telegram-bot==21.3
groq==0.9.0
gspread==6.1.2
google-auth==2.29.0
python-dotenv==1.0.1
httpx==0.27.0
